refactor(state): move reducer into its own file
